Moved Test, TestSuite and TestsCollection to another folder
